requirements.txt
python-telegram-bot[job-queue]==20.4
httpx==0.24.1
aiofiles==23.1.0
python-dotenv